# backend/internal/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Arrflix API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/downloaders": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "downloaders"
                ],
                "summary": "List downloaders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dbgen.Downloader"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "downloaders"
                ],
                "summary": "Create downloader",
                "parameters": [
                    {
                        "description": "Create downloader",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.DownloaderCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dbgen.Downloader"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/downloaders/test": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "downloaders"
                ],
                "summary": "Test downloader connection with form configuration",
                "parameters": [
                    {
                        "description": "Test downloader configuration",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.DownloaderTestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/downloader.TestResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/downloaders/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "downloaders"
                ],
                "summary": "Get downloader",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dbgen.Downloader"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "downloaders"
                ],
                "summary": "Update downloader",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Downloader ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update downloader",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.DownloaderUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dbgen.Downloader"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "downloaders"
                ],
                "summary": "Delete downloader",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Downloader ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/downloaders/{id}/test": {
            "post": {
                "tags": [
                    "downloaders"
                ],
                "summary": "Test downloader connection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Downloader ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/downloader.TestResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dbgen.Downloader": {
            "type": "object",
            "required": [
                "config_json",
                "created_at",
                "default",
                "enabled",
                "id",
                "name",
                "password",
                "protocol",
                "type",
                "updated_at",
                "url",
                "username"
            ],
            "properties": {
                "config_json": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "default": {
                    "type": "boolean"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "downloader.TestResult": {
            "type": "object",
            "required": [
                "message",
                "success"
            ],
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "version": {
                    "type": "string"
                },
                "webApiVersion": {
                    "type": "string"
                }
            }
        },
...
        "handlers.DownloaderCreateRequest": {
            "type": "object",
            "required": [
                "config_json",
                "default",
                "enabled",
                "name",
                "password",
                "protocol",
                "type",
                "url",
                "username"
            ],
            "properties": {
                "config_json": {
                    "type": "object",
                    "additionalProperties": true
                },
                "default": {
                    "type": "boolean"
                },
                "enabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "handlers.DownloaderTestRequest": {
            "type": "object",
            "required": [
                "config_json",
                "password",
                "type",
                "url",
                "username"
            ],
            "properties": {
                "config_json": {
                    "type": "object",
                    "additionalProperties": true
                },
                "password": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "handlers.DownloaderUpdateRequest": {
            "type": "object",
            "required": [
                "config_json",
                "default",
                "enabled",
                "name",
                "password",
                "protocol",
                "type",
                "url",
                "username"
            ],
            "properties": {
                "config_json": {
                    "type": "object",
                    "additionalProperties": true
                },
                "default": {
                    "type": "boolean"
                },
                "enabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}